refactor: 상품/재고 서비스 Kafka 수정 및 설정 변경
