+++ client/product/src/main/java/com/live_commerce/product/inventory/infrastructure/config/KafkaConsumerConfig.java
@@ -1,7 +1,9 @@
 package com.live_commerce.product.inventory.infrastructure.config;
 
+import lombok.RequiredArgsConstructor;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
+import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.annotation.EnableKafka;
@@ -17,40 +19,36 @@ import java.util.Map;
 
 @EnableKafka
 @Configuration
+@RequiredArgsConstructor
 public class KafkaConsumerConfig {
 
-    // Kafka 컨슈머 팩토리를 생성하는 빈을 정의합니다.
-    // ConsumerFactory는 Kafka 컨슈머 인스턴스를 생성하는 데 사용됩니다.
-    // 각 컨슈머는 이 팩토리를 통해 생성된 설정을 기반으로 작동합니다.
+    private final KafkaProperties kafkaProperties;
+
     @Bean
-    public ConsumerFactory<String, Object> consumerFactory() {
-        JsonDeserializer<Object> jsonDeserializer = new JsonDeserializer<>();
-        jsonDeserializer.addTrustedPackages("*");
-
-        // 컨슈머 팩토리 설정을 위한 맵을 생성합니다.
-        Map<String, Object> configProps = new HashMap<>();
-        configProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
+    public ConsumerFactory<String, Object> inventoryConsumerFactory() {
+        Map<String, Object> configProps = new HashMap<>(kafkaProperties.buildConsumerProperties());
         configProps.put(ConsumerConfig.GROUP_ID_CONFIG, "inventory-group");
-
-        configProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
-        configProps.put(ErrorHandlingDeserializer.KEY_DESERIALIZER_CLASS, StringDeserializer.class);
-
-        configProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
-        configProps.put(ErrorHandlingDeserializer.VALUE_DESERIALIZER_CLASS, JsonDeserializer.class);
-
-        return new DefaultKafkaConsumerFactory<>(configProps, new StringDeserializer(), jsonDeserializer);
+        return new DefaultKafkaConsumerFactory<>(configProps);
     }
 
-    // Kafka 리스너 컨테이너 팩토리를 생성하는 빈을 정의합니다.
-    // ConcurrentKafkaListenerContainerFactory는 Kafka 메시지를 비동기적으로 수신하는 리스너 컨테이너를 생성하는 데 사용됩니다.
-    // 이 팩토리는 @KafkaListener 어노테이션이 붙은 메서드들을 실행할 컨테이너를 제공합니다.
     @Bean
     public ConcurrentKafkaListenerContainerFactory<String, Object> kafkaListenerContainerFactory() {
-        // ConcurrentKafkaListenerContainerFactory를 생성합니다.
         ConcurrentKafkaListenerContainerFactory<String, Object> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        // 컨슈머 팩토리를 리스너 컨테이너 팩토리에 설정합니다.
-        factory.setConsumerFactory(consumerFactory());
-        // 설정된 리스너 컨테이너 팩토리를 반환합니다.
+        factory.setConsumerFactory(inventoryConsumerFactory());
+        return factory;
+    }
+
+    @Bean
+    public ConsumerFactory<String, Object> productConsumerFactory() {
+        Map<String, Object> configProps = new HashMap<>(kafkaProperties.buildConsumerProperties());
+        configProps.put(ConsumerConfig.GROUP_ID_CONFIG, "product-group");
+        return new DefaultKafkaConsumerFactory<>(configProps);
+    }
+
+    @Bean
+    public ConcurrentKafkaListenerContainerFactory<String, Object> productKafkaListenerContainerFactory() {
+        ConcurrentKafkaListenerContainerFactory<String, Object> factory = new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(productConsumerFactory());
         return factory;
     }
 }
