+++ client/product/src/main/java/com/live_commerce/product/inventory/infrastructure/kafka/consumer/InventoryEventConsumer.java
@@ -1,8 +1,9 @@
-package com.live_commerce.product.inventory.application.service;
+package com.live_commerce.product.inventory.infrastructure.kafka.consumer;
 
-import com.live_commerce.product.inventory.application.dto.event.InventoryDecreasedEvent;
-import com.live_commerce.product.inventory.application.dto.event.InventoryFailedEvent;
-import com.live_commerce.product.inventory.application.dto.event.OrderCreatedEvent;
+import com.live_commerce.product.inventory.application.service.InventoryService;
+import com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryDecreasedEvent;
+import com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryFailedEvent;
+import com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent;
 import com.live_commerce.product.inventory.domain.exception.InventoryException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,12 +14,12 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class InventoryConsumer {
+public class InventoryEventConsumer {
 
     private final InventoryService inventoryService;
     private final KafkaTemplate<String, Object> kafkaTemplate;
 
-    @KafkaListener(topics = "order-created", groupId = "inventory-group", containerFactory = "kafkaListenerContainerFactory")
+    @KafkaListener(topics = "order-created", containerFactory = "inventoryKafkaListenerContainerFactory")
     public void consumeOrderCreated(OrderCreatedEvent event) {
         log.info("order-created 이벤트 수신: {}", event);
 
@@ -33,16 +34,16 @@ public class InventoryConsumer {
             kafkaTemplate.send("inventory-decreased", decreasedEvent);
             log.info("inventory-decreased 이벤트 발행 완료: {}", decreasedEvent);
         } catch (InventoryException e) {
-            log.error("재고 차감 실패: {}", e.getMessage());
-            // 실패 이벤트
-            InventoryFailedEvent failedEvent = new InventoryFailedEvent(
-                    event.orderId(),
-                    event.productId(),
-                    event.quantity(),
-                    e.getMessage()
-            );
-            kafkaTemplate.send("inventory-failed", failedEvent);
-            log.info("inventory-failed 이벤트 발행 완료: {}", failedEvent);
+            log.error("재고 차감 실패: {}, 이유: {}", event.orderId(), e.getMessage());
+
+//            InventoryFailedEvent failedEvent = new InventoryFailedEvent(
+//                    event.orderId(),
+//                    event.productId(),
+//                    event.quantity(),
+//                    e.getMessage()
+//            );
+//            kafkaTemplate.send("inventory-failed", failedEvent);
+//            log.info("inventory-failed 이벤트 발행 완료: {}", failedEvent);
         }
     }
