+++ client/product/src/main/java/com/live_commerce/product/product/presentation/controller/ProductController.java
@@ -1,6 +1,6 @@
 package com.live_commerce.product.product.presentation.controller;
 
-import com.live_commerce.product.inventory.application.service.ProductRankingService;
+import com.live_commerce.product.product.application.service.ProductRankingService;
 import com.live_commerce.product.product.application.dto.*;
 import com.live_commerce.product.product.application.service.ProductService;
 import com.live_commerce.product.product.infrastructure.common.ResponseUtil;
