+++ server/config/src/main/resources/config-repo/product-dev.yml
@@ -32,10 +32,30 @@ spring:
   redis:
     host: localhost
     port: 6379
-  kafka:
-    consumer:
-      reconnect-backoff-ms: 5000
-      reconnect-backoff-max-ms: 30000
+
+    kafka:
+      bootstrap-servers: localhost:9092
+      producer:
+        key-serializer: org.apache.kafka.common.serialization.StringSerializer
+        value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+        properties:
+          spring.json.type.mapping: >
+            order-created:com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent,
+            inventory-decreased:com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryDecreasedEvent
+
+      consumer:
+        auto-offset-reset: earliest
+        key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+        value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
+        properties:
+          spring.json.trusted.packages: com.live_commerce
+          spring.json.type.mapping: >
+            order-created:com.live_commerce.product.inventory.infrastructure.kafka.event.OrderCreatedEvent,
+            inventory-decreased:com.live_commerce.product.inventory.infrastructure.kafka.event.InventoryDecreasedEvent
+        listener:
+          ack-mode: MANUAL_IMMEDIATE
+        #      reconnect-backoff-ms: 5000
+        #      reconnect-backoff-max-ms: 30000
 
 gateway:
   base-url: http://localhost:19091
